requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cx-Freeze==6.7
cx-Logging==3.0
et-xmlfile==1.1.0
fpdf==1.7.2
idna==3.2
importlib-metadata==4.6.1
multidict==5.1.0
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
Pillow==8.3.1
psutil==5.8.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
reportlab==3.5.68
six==1.16.0
tkfilebrowser==2.3.2
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
yarl==1.6.3
zipp==3.5.0
